requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiodns==2.0.0             # via -r requirements.in
aiohttp==3.6.2            # via -r requirements.in, aiojobs, magnet2torrent
aiojobs==0.2.2            # via -r requirements.in
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp
bencode.py==3.0.1         # via -r requirements.in
cchardet==2.1.6           # via -r requirements.in
cffi==1.14.0              # via pycares
chardet==3.0.4            # via aiohttp
expiringdict==1.2.1       # via magnet2torrent
idna==2.9                 # via yarl
lxml==4.5.1               # via -r requirements.in
magnet2torrent==1.1.0     # via -r requirements.in
multidict==4.7.6          # via aiohttp, yarl
pycares==3.1.1            # via aiodns
pycparser==2.20           # via cffi
six==1.15.0               # via magnet2torrent
tabulate==0.8.7           # via -r requirements.in
tqdm==4.46.1              # via -r requirements.in
yarl==1.4.2               # via aiohttp
